Inclusion of protonmail-bridge resolves T8937
Summary:
Replaces D9816 due to lack of progress.

I have a ProtonMail subscription and would like to maintain this.

ProtonMail Bridge requires either `gnome-keyring` or `pass` but will default to `pass` if both are installed. Should I add a rundep on either?

Test Plan: Installed and configured the bridge. Works flawlessly with Thunderbird.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D10348