Update to 1.8.7 and enable updates
Summary:
Along with updating to 1.8.7 this enables the auto update feature. New updates will be installed to the users home directory. The launcher `protonmail-bridge` determines whether to run the default `proton-bridge` or a newer version if present. In the future this will allow hotfix releases such as 1.8.1 and 1.8.2 to be delivered to the users much faster. Additionally, users can now update to early access releases if they wish to.

**Summarized changelog:**

 - 1.8.7
  - Updated golang Secure Remote Password Protocol
  - Updated crypto-libraries to gopenpgp/v2 v2.1.10
  - Improved parsing of message headers
  - Sync and performance improvements
 - 1.8.3
  - Fixed a bug with sending encrypted emails to external contacts
 - 1.8.2
  - Hotfix for error during bug reporting
 - 1.8.1
  - Hotfix for crash when listing empty folder
 - 1.8.0
  - Refactor of message builder to achieve greater RFC compliance
  - Implemented connection manager to improve performance during weak connection, better handling of connection loss and other connectivity issues
  - Increased the number of message fetchers to allow more parallel requests - performance improvement
  - Log changes for easier debugging (update-related)
  - Fixed bugs with building of message bodies/headers
  - Fixed incorrect naming format of some of the attachments
  - Removed html-wrapping of non-decriptable messages - to facilitate decryption outside Bridge and/or allow to store such messages as they are

Full changelog: https://protonmail.com/download/bridge/stable_releases.html

Test Plan:
Synced and sent mail through Thunderbird successfully.

Successfully used the update feature to update to early access and downgrade to 1.8.7 again.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: algent, JoshStrobl

Differential Revision: https://dev.getsol.us/D11182